Long stories
Most of these are alternative names. Some of them are original names as determined by the people who discovered them. The rest were temporary names that you can still find on a lot of periodic tables. That second category exists because there was a need for a name but an uncontroversial one was needed until a name could be determined.